pyproject.toml
[tool.poetry]
name = "eigen"
version = "0.1.0"
description = "Backend engineering challenge."
authors = ["Pedro Allen Revez"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.0"
rich = "^13.3.5"
nltk = "^3.8.1"
dagster = "^1.3.6"
scikit-learn = "^1.2.2"
dagit = "^1.3.6"
typer = "^0.7.0"
spacy = "3.5.3"
python-levenshtein = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.12.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
eigen = "eigen.cli:main"

[tool.dagster]
module_name = "eigen"
